Merge AllocationInfo in C and C++ API
Hide AllocationInfo struct details from the public headers (A request from Shah Asaduzzaman)

- rename onnxruntime::AllocatorInfo to ::ONNXRuntimeAllocatorInfo
- rename onnxruntime::AllocatorType to ::ONNXRuntimeAllocatorType
- rename onnxruntime::MemType to ::ONNXRuntimeMemType

Related work items: #1769

+++ onnxruntime/core/graph/initializer.h
@@ -319,7 +319,7 @@ class Initializer final {
 
   inline void scale_by_axis(const Initializer& other, int axis) {
     int64_t num = 1;
-    for (int k = axis; k < dims_.size(); k++) {
+    for (size_t k = axis; k < dims_.size(); k++) {
       num *= dims_[k];
     }
