+++ onnxruntime/test/custom_op_shared_lib/test_custom_op.cc
@@ -23,7 +23,7 @@ class FooKernel : public OpKernel {
     Tensor* Y = ctx->Output(0, X->Shape());
     auto* Y_data = Y->template MutableData<float>();
 
-    for (size_t i = 0; i < X->Shape().Size(); i++) {
+    for (int64_t i = 0; i < X->Shape().Size(); i++) {
       Y_data[i] = X_data[i] + W_data[i];
     }
